一覧画面にSNS設定済かどうか把握できるようにアイコン表示
