+++ app/controllers/admin/breweries_controller.rb
@@ -1,5 +1,6 @@
 class Admin::BreweriesController < AdminController
   include BrewerySearchModule
+  include SocialAccountSearchModule
   before_action :set_brewery, only: [:edit, :destroy, :update]
   before_action :revise_params, only: [:create, :update]
   before_action :generate_social_account_list, only: [:new, :edit]
@@ -49,16 +50,6 @@ class Admin::BreweriesController < AdminController
     params.require(:brewery).permit(:name, :web_site, :image,
       social_accounts_attributes: [:id, :account_type, :url, :comment]
     )
-  end
-
-  def generate_social_account_list
-    # {"facebook"=>0, "twitter"=>1, "hatena"=>2・・という構造から
-    # [{:index=>0, :name=>"facebook"}, {:index=>1, :name=>"twitter"}という構造に変更するために
-    @social_accounts =[]
-    SocialAccount.account_types.keys.each_with_index do|account, index|
-      @social_accounts.push({index: index, name: account })
-    end
-    @social_accounts
   end
 
   def revise_params
