+++ app/models/place.rb
@@ -9,7 +9,8 @@ class Place < ActiveRecord::Base
   has_many :items, through: :place_item_relays
   has_many :place_social_account_relays, dependent: :destroy
   has_many :social_accounts, through: :place_social_account_relays
-
+  
+  accepts_nested_attributes_for :social_accounts
   def self.place_type_list
     result = []
     self.place_types.keys.each_with_index do|account, index|
