+++ app/views/admin/places/_form.html.erb
@@ -54,7 +54,7 @@
         
       </div>      
     </div>
-
+    <%= f.hidden_field :social_accounts, value: "{{ selectedSocialAccounts }}" %>
   </div>  
 </div>
 <script src="https://maps.googleapis.com/maps/api/js"></script>
