+++ app/views/admin/places/edit.html.erb
@@ -2,9 +2,12 @@
   <div class="col-xs-12">
     <%= form_for(@place, url: admin_place_path, html: { class: 'form-horizontal'}) do |f| %>
       <%= render partial: 'form' , locals: {f: f} %>
+      <div class="col-xs-10">
+        <social-account-list data="<%= @place_social_accounts.to_json %>" accounts="selectedSocialAccounts" attributes="<%= @social_accounts.to_json %>" ></social-account-list>
+      </div>
       <div class="row">
         <div class="col-md-2 col-md-offset-4">
-          <ul class="list-inline">
+          <ul class="list-inline">            
             <li><%= link_to '戻る', admin_places_path, class: 'btn btn-default' %></li>
             <li><%= f.submit '更新する', class: 'btn btn-primary'%><li>
           </ul>
